cleanup

+++ sperl_heap.c
@@ -6,7 +6,7 @@
 
 void* SPerl_HEAP_alloc(SPerl* sperl, int64_t size) {
   (void) sperl;
-  if (size < 0 || SIZE_MAX < (int64_t) size) {
+  if (size < 0 || INT64_MAX < (int64_t) size) {
     SPerl_ALLOCATOR_UTIL_exit_with_malloc_failure();
   }
   return SPerl_ALLOCATOR_UTIL_safe_malloc(size, sizeof(uint8_t));
